meeting/250530/requirements_for_mac.txt
ffmpeg-python==0.2.0
numpy==1.23.5
imageio==2.31.1
tensorflow-macos==2.11.0
tensorflow-metal==0.7.0
keras==2.11.0
glob2==0.7
argparse==1.4.0
pandas==1.5.3
tqdm==4.66.1
scipy
scikit-learn